Update release script

+++ dev-bin/release.sh
@@ -60,6 +60,15 @@ if [ "$should_continue" != "y" ]; then
     exit 1
 fi
 
+mvn test
+
+read -r -n 1 -p "Continue given above tests? (y/n) " should_continue
+
+if [ "$should_continue" != "y" ]; then
+    echo "Aborting"
+    exit 1
+fi
+
 page=.gh-pages/index.md
 cat <<EOF > $page
 ---
@@ -71,30 +80,29 @@ version: $tag
 
 EOF
 
+mvn versions:set -DnewVersion="$version"
+
 perl -pi -e "s/(?<=<version>)[^<]*/$version/" README.md
 perl -pi -e "s/(?<=com\.maxmind\.geoip2\:geoip2\:)\d+\.\d+\.\d+([\w\-]+)?/$version/" README.md
 
 cat README.md >> $page
 
-if [ -n "$(git status --porcelain)" ]; then
-    git diff
+git diff
 
-    read -r -n 1 -p "Commit README.md changes? " should_commit
-    if [ "$should_commit" != "y" ]; then
-        echo "Aborting"
-        exit 1
-    fi
-    git add README.md
-    git commit -m 'update version number in README.md'
+read -r -n 1 -p "Commit changes? " should_commit
+if [ "$should_commit" != "y" ]; then
+    echo "Aborting"
+    exit 1
 fi
+git add README.md pom.xml
+git commit -m "Preparing for $version"
 
-# could be combined with the primary build
-mvn release:clean
-mvn release:prepare -DreleaseVersion="$version" -Dtag="$tag"
-mvn release:perform
+mvn clean deploy
+
 rm -fr ".gh-pages/doc/$tag"
-cp -r target/checkout/target/reports/apidocs ".gh-pages/doc/$tag"
-ln -Tfs "$tag" .gh-pages/doc/latest
+cp -r target/reports/apidocs ".gh-pages/doc/$tag"
+rm .gh-pages/doc/latest
+ln -fs "$tag" .gh-pages/doc/latest
 
 pushd .gh-pages
 
@@ -106,7 +114,6 @@ echo "Release notes for $version:
 $notes
 
 "
-
 read -r -n 1 -p "Push to origin? " should_push
 
 if [ "$should_push" != "y" ]; then
@@ -119,8 +126,7 @@ git push
 popd
 
 git push
-git push --tags
 
 gh release create --target "$(git branch --show-current)" -t "$version" -n "$notes" "$tag" \
-    "target/checkout/target/geoip2-$version-with-dependencies.zip" \
-    "target/checkout/target/geoip2-$version-with-dependencies.zip.asc"
+    "target/geoip2-$version-with-dependencies.zip" \
+    "target/geoip2-$version-with-dependencies.zip.asc"
